src/gausskernel/dbmind/dbmind/requirements-x86.txt
## For DBMind-core ##
paramiko>=2.7.2
numpy
scikit-learn
requests
sql-metadata
sqlparse
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pycryptodome
## For components ##
mo-sql-parsing
prettytable>=2.5.0
## Prometheus Exporter
pyyaml
prometheus-client
## X-Tuner ##
bayesian-optimization
